requirements.txt
lark-oapi==1.5.3
anthropic>=0.40.0
openai>=1.0.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
pymysql>=1.1.0
httpx>=0.27.0
paramiko>=4.0.0
